+++ src/Core/Query/AbstractResponseParser.php
@@ -51,30 +51,6 @@ abstract class AbstractResponseParser
     }
 
     /**
-     * Parses header data (if available) and adds it to result data.
-     *
-     * @param array $data
-     * @param array $result
-     *
-     * @return array
-     */
-    public function addHeaderInfo(array $data, array $result): array
-    {
-        $status = null;
-        $queryTime = null;
-
-        if (isset($data['responseHeader'])) {
-            $status = $data['responseHeader']['status'];
-            $queryTime = $data['responseHeader']['QTime'];
-        }
-
-        $result['status'] = $status;
-        $result['queryTime'] = $queryTime;
-
-        return $result;
-    }
-
-    /**
      * Parses HTTP status code and adds boolean wasSuccessful to result data.
      * Parses HTTP status message and adds string statusMessage to result data.
      *