+++ src/QueryType/Luke/ResponseParser/Index.php
@@ -38,8 +38,6 @@ class Index extends AbstractResponseParser implements ResponseParserInterface
         $data['fieldsResult'] = null;
         $data['infoResult'] = null;
 
-        $data = $this->addHeaderInfo($data, $data);
-
         return $data;
     }