+++ src/QueryType/Server/Collections/ResponseParser/ClusterStatus.php
@@ -29,8 +29,6 @@ class ClusterStatus extends ResponseParser
     {
         $data = $result->getData();
 
-        $data = $this->addHeaderInfo($data, $data);
-
         return $data;
     }
 }