+++ tests/QueryType/Analysis/Result/FieldTest.php
@@ -64,7 +64,6 @@ class FieldDummy extends Field
     public function __construct($status, $queryTime, $items)
     {
         $this->items = $items;
-        $this->queryTime = $queryTime;
-        $this->status = $status;
+        $this->responseHeader = ['status' => $status, 'QTime' => $queryTime];
     }
 }